Add files via upload
The following code is written in Python to create a simple, rule-based chatbot. With the booming craze of AI chatbots like Chat GPT or Bard in the market nowadays, I decided to try making one myself. Despite this being the very basic step, I found it fascinating to learn about all the efforts that go into designing and training the chatbots to near perfection. I'm looking forward to learning more.